Item service impl completion and bug fixes related to item related components.

+++ src/main/java/com/ijse/lostandfoundsystem/service/ItemService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ItemService {
     void reportItem(ItemDTO itemDTO);
-    ItemDTO getSelectedItem(Long itemId);
+    ItemDTO getSelectedItem(int itemId);
     List<ItemDTO> getAllItems();
     void updateItem(Long itemId, ItemDTO itemDTO);
     void deleteItem(Long itemId);
